requirements.txt
python-telegram-bot>=21.6,<22
cloudscraper>=1.2
python-telegram-bot[job-queue]
python-telegram-bot[rate-limiter]
google-genai>=1.47.0
